pyproject.toml
[tool.poetry]
name = "zadanie"
version = "0.1.0"
description = ""
authors = ["ByteMyPawn <michael.mikolaj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
openai = "^1.55.3"
fastapi = "^0.115.5"
python-dotenv = "^1.0.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.18"
jwt = "^1.3.1"
bcrypt = "^4.2.1"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.32.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
